Внешний API и внутренняя реализация TaskController разделены - TaskController теперь интерфейс, а реализация лежит в AbstractTaskController
